Add files via upload
This is the key file for the segmentation. We are using Encoder-Decoder model for the segmentation. The Encoder block extracts the features of the tumor region of the brain from the fused images and the decoder block converts the feature into visible image, which represents the tumor, segmented from the fused MRI and CT images.